# src/main/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: Report API
  description: API for managing reports
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
paths:
  /reports:
    get:
      summary: Get all reports
      operationId: getAllReports
      responses:
        '200':
          description: A list of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportResponse'
    post:
      summary: Create a new report
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
      responses:
        '200':
          description: The created report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /reports/{id}:
    get:
      summary: Get a report by ID
      operationId: getReportById
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the report to retrieve
      responses:
        '200':
          description: The requested report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a report
      operationId: updateReport
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the report to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
      responses:
        '200':
          description: The updated report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a report
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the report to delete
      responses:
        '204':
          description: Report deleted
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ReportCreateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required:
        - title
        - description
    ReportResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
    ErrorResponse:
      type: object
      properties:
        message:
          type: string